Tracker: Improve repeated apps in Open with... menu
If you have multiple Haiku installations mounted, it is common to
have the same app on multiple volumes. This makes the Open with...
menu show the full paths truncating them which makes it tough to
see the app name you're looking for.

Show volume and app names for repeated apps in Open with... menu
instead of showing the full path. Keep track of whether each app in
the list is a repeat and only display the volume for those.

If there is only 1 app name, print it.
If >1 app name and volume names are different, print
    [volume name] app name
If >1 app name but volume names are the same, print full path

Because many apps are in packages the volume is listed as [system]

Make kMaxMenuWidth a bit wider.

Compare app names and volume names using locale collator.

Sort by name, determine labels, then sort by relation because
repeats are not always consecutive once sorted by relation.

Change-Id: I2c6c1db99065641cf4f940a69a47693480f151aa
Reviewed-on: https://review.haiku-os.org/c/haiku/+/5448
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>
Reviewed-by: Adrien Destugues <pulkomandy@pulkomandy.tk>