pkgman: change default search scope
* Reduce default search scope to: "name", "summary", and "provides".
* Add a new "--search-scope=<scope>" option that allows to either limit
  the search to packages names (-s name), or to replicate the previous
  behavior of searching everywhere in the package info (-s full).

Motivation for the change:

The ever increasing number of packages available for Haiku was impacting
pkgman search usefulness, as it easily returns too many packages (what a
wonderful problem to have :-D).

Limiting the default search scope helps with that, and the old behavior
is kept "just an option away".

Change-Id: I5b456b90137237134eee7ca7ee501bf8e3767440
Reviewed-on: https://review.haiku-os.org/c/haiku/+/5616
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>
Reviewed-by: Jérôme Duval <jerome.duval@gmail.com>