build: clean HdsSchema directories
Previously, HdsSchema directories were "built" with the Clean action and
manually created in their respective commands. This patch creates the
directory as a separate target to be cleaned. The MkDir1 action is used
directly since $(1) is already rooted and the HdsSchema actions use the
ungristed directory as a dependency.

This also fixes a build failure in Ham since it does not allow multiple
targets as part of 'together' actions.

Change-Id: Ibc818375477b5f1712ac7c7cec8a1f8812f62b65
Reviewed-on: https://review.haiku-os.org/c/haiku/+/5630
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>
Reviewed-by: Adrien Destugues <pulkomandy@pulkomandy.tk>
Reviewed-by: Stephan Aßmus <superstippi@gmx.de>