AboutSystem: update CPU info on replicant constructor.
This is similar to changes made in hrev56526 and hrev56815,
but for CPU related info this time.

Extracted some new _Get*() methods for the code that will get
called from both the normal and the replicant constructors
of SysInfoView.

Re-arranged the code a bit so the BStringViews get created /
updated in the order they appear on the GUI.

Change-Id: I2a371ec623e0b4d7497056fae6d5346b98d07ee2
Reviewed-on: https://review.haiku-os.org/c/haiku/+/6396
Reviewed-by: John Scipione <jscipione@gmail.com>
Reviewed-by: waddlesplash <waddlesplash@gmail.com>